IA review: change volume to Q8.8 management
remove vol*vol quadratic management (db already quadratic)

+++ Core/USB_DEVICE/App/ak4490r.h
@@ -129,12 +129,12 @@
 	uint8_t control8;
 } AK4490R_RegisterTypeDef;
 
-extern volatile uint8_t configured_volume;
+extern volatile int16_t configured_volume;
 extern volatile bool configured_mute;
 extern volatile bool audio_stop_pending;
 
 uint8_t AK4490R_DAC_Init();
-uint8_t AK4490R_DAC_SetVolume(uint8_t vol);
+uint8_t AK4490R_DAC_SetVolume(int16_t vol);
 uint8_t AK4490R_DAC_SetMute(uint8_t mute);
 void AK4490R_DAC_SetMute_Force(void);
 uint8_t AK4490R_DAC_SetFormat(uint8_t format);
